+++ test/dummy/rails60_webpacker4/app/models/ship.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class Ship < ApplicationRecord
-  has_many :members
-end