+++ test/dummy/rails61_webpacker5/app/models/favorite_thing.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class FavoriteThing < ApplicationRecord
-  belongs_to :member
-
-  validates :name, presence: true
-
-  accepts_nested_attributes_for :member
-end