+++ test/dummy/rails61_webpacker5/config/initializers/super.rb
@@ -1,3 +1,0 @@
-Super.configuration do |c|
-  c.title = "My Admin Site"
-end