+++ test/dummy/rails61_webpacker5/db/seeds.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require "super_test_engine/starfleet_seeder"
-StarfleetSeeder.seed