+++ test/dummy/rails61_webpacker6/config/initializers/super.rb
@@ -1,5 +1,0 @@
-Super.configuration do |c|
-  c.title = "My Admin Site"
-  c.javascripts = Super::Assets.use_webpacker(c.javascripts)
-  c.stylesheets = Super::Assets.use_webpacker(c.stylesheets)
-end