+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Speed up rendering `Super::Link`
 * Speed up rendering `Super::ViewChain`
 
+### Deprecations
+
+* Stopped supporting Webpacker
+
 ## [0.21.0] - 2022-06-05
 
 ### Improvements